requirements.txt
absl-py==0.7.1
alibi==0.5.1
arff2pandas==1.0.1
astor==0.7.1
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.0
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.12.3
changefinder==0.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorful==0.5.4
cycler==0.10.0
cymem==2.0.3
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
gast==0.2.2
gdown==3.10.3
google-pasta==0.2.0
grpcio==1.16.1
h5py==2.8.0
idna==2.9
imageio==2.9.0
importlib-metadata==1.7.0
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.1.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kmeans-pytorch==0.3
liac-arff==2.4.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.0.1
metakernel==0.24.4
mistune==0.8.4
mne==0.20.7
mock==3.0.5
more-itertools==7.0.0
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.4
nose==1.3.7
notebook==5.7.8
octave-kernel==0.32.0
olefile==0.46
opencv-python==4.3.0.36
openpyxl==3.0.7
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.2.0
plac==1.1.3
plotly==4.5.4
plotly-express==0.4.1
preshed==3.0.2
prettyprinter==0.18.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
pycparser==2.19
pydicom==2.1.2
Pygments==2.4.2
pylibjpeg==1.3.0
pylibjpeg-openjpeg==1.1.1
pylibjpeg-rle==1.1.0
pyparsing==2.4.0
pyreadline==2.1
pyrsistent==0.14.11
PySocks==1.7.1
python-dateutil==2.8.0
python-gdcm==3.0.9.0
pytz==2019.1
PyWavelets==1.1.1
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.0
qtconsole==4.5.1
requests==2.23.0
retrying==1.3.3
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scikit-learn==0.21.2
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
shap==0.35.0
six==1.12.0
sklearn==0.0
soupsieve==2.0.1
spacy==2.3.2
spacy-lookups-data==0.3.2
srsly==1.0.2
statsmodels==0.10.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
Theano==1.0.4
thinc==7.4.1
threadpoolctl==2.1.0
tornado==6.0.2
traitlets==4.3.2
typing-extensions==3.7.4.2
urllib3==1.25.9
wasabi==0.7.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wfdb==3.1.0
wget==3.2
widgetsnbextension==3.4.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
numpy==1.19.5
tifffile==2020.7.4
attrs==19.1.0
GPUtil==1.4.0